Move work data in det_manip to work_data.hpp
